# docs/swagger.yaml
definitions:
...
  models.File:
    properties:
      bucket_id:
        type: string
      checksum:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      file_size:
        type: integer
      file_size_limit:
        type: integer
      finalized:
        type: boolean
      id:
        type: string
      metadata:
        type: string
      name:
        type: string
      path:
        type: string
      updated_at:
        type: string
    required:
    - bucket_id
    - name
    type: object
...
  router.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: error message
        type: string
    type: object
  router.InitiateFileUploadDTO:
    properties:
      bucketCode:
        type: string
      fileSizeLimit:
        type: integer
      regionId:
        type: string
    required:
    - regionId
    type: object
  router.InitiateFileUploadResponse:
    properties:
      ttl:
        description: seconds
        type: integer
      url:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/file/:
    post:
      consumes:
      - application/json
      description: Initiate a new file upload. Receives regionId and bucketCode, returns
        a pre-signed upload URL and TTL (seconds). Admin access required.
      operationId: InitiateFileUpload
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Upload initiation data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/router.InitiateFileUploadDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/router.InitiateFileUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Initiate file upload
      tags:
      - files
  /api/v1/file/{blob}/finalize:
    post:
      description: Finalize a file upload after client notifies server. Admin access
        required.
      operationId: FinalizeFileUpload
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Blob ID
        in: path
        name: blob
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.File'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Finalize file upload
      tags:
      - files
  /api/v1/file/{id}:
    delete:
      description: Delete a file by ID. Admin access required.
      operationId: DeleteFile
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Delete file
      tags:
      - files
...
  /api/v1/upload/{blob}:
    patch:
      consumes:
      - application/octet-stream
      - multipart/form-data
      - application/x-www-form-urlencoded
      description: Upload file data using the blob ID provided by the server. Supports
        stream, form-data, and multipart uploads. No admin access required.
      operationId: UploadFileBlob
      parameters:
      - description: API Token
        in: header
        name: x-api-token
        required: true
        type: string
      - description: Blob ID
        in: path
        name: blob
        required: true
        type: string
      - description: File data (multipart or form-data, required if not using raw
          stream)
        in: formData
        name: file
        type: file
      - description: File data (raw stream, required if not using multipart/form-data)
        in: body
        name: file
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Upload file data
      tags:
      - upload
swagger: "2.0"